% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_missing.R
\name{explore_missing}
\alias{explore_missing}
\title{Shows exactly where there is missing data and how much data is missing.}
\usage{
explore_missing(data)
}
\arguments{
\item{data}{A data.frame: The target dataframe to explore}

\item{num_rows}{integer: The number of rows above and below the missing value to output}
}
\value{
dataframe: The resultant dataframes
}
\description{
Returns 2 dataframes: 1 dataframe of exactly where there is
missing data and another table how much data is missing for each feature.
}
\examples{
explore_missing(iris)
}
