% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_outliers.R
\name{explore_outliers}
\alias{explore_outliers}
\title{Explores outliers in each feature of dataset based on standard deviation.}
\usage{
explore_outliers(data)
}
\arguments{
\item{data}{A data.frame: The target dataframe to explore}
}
\value{
data.frame: Dataframe with details on the outliers
}
\description{
Explores outliers in each feature of dataset based on standard deviation.
}
\examples{
explore_outliers(iris)
}
