% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_summary.R
\name{explore_summary}
\alias{explore_summary}
\title{Provides basic summary statistics and wholistic
information about all the features of provided data}
\usage{
explore_summary(df)
}
\arguments{
\item{df}{dataframe (clean)}
}
\value{
a summary dataframe
}
\description{
Provides basic summary statistics and wholistic
information about all the features of provided data
}
\examples{
explore_summary(iris)
}
