% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_featureviz.R
\name{get_featureviz}
\alias{get_featureviz}
\title{Gets correlation plots of features in data (NA)}
\usage{
get_featureviz(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
plot
}
\description{
Gets correlation plots of features in data (NA)
}
\examples{
get_featureviz(iris)
}
