% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers.R
\name{get_outliers}
\alias{get_outliers}
\title{Gets data outliers}
\usage{
get_outliers(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
dataframe
}
\description{
Gets data outliers
}
\examples{
get_outliers(iris)
}
