% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Gets data summary (NA)}
\usage{
get_summary(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
dataframe
}
\description{
Gets data summary (NA)
}
\examples{
get_summary(iris)
}
