% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_feature_map.R
\name{explore_feature_map}
\alias{explore_feature_map}
\title{Returns a cumulative faceted plot on pairwise feature relationships.}
\usage{
explore_feature_map(df)
}
\arguments{
\item{data}{A data.frame: The target dataframe to explore}
}
\value{
plot
}
\description{
The plot consists of NxN mini-charts where N is number of features.
Main diagonal shows feature distribution. Pairwise Pearson correlations
are shown above main diagonal. Pairwise feature joint distributions
are shown below main diagonal.
}
\details{
Note: non-numeric features will be skipped!
}
\examples{
explore_feature_map(iris)
}
