% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_missing.R
\name{explore_missing}
\alias{explore_missing}
\title{Shows exactly where there is missing data and how much data is missing.}
\usage{
explore_missing(data, num_rows = 0, type = "location")
}
\arguments{
\item{data}{A data.frame: The target dataframe to explore}

\item{num_rows}{integer: The number of rows above and below the missing value to output}

\item{type}{character: The desired type of output (location or count)}
}
\value{
dataframe: The resultant dataframes
}
\description{
explore_missing will explore missing observations within data. It will
return 1 of 2 tables: 1 table of exactly where there is missing data or
another table showing how many observations are missing, and the
proportion of how much data is missing for each feature.
}
\examples{
explore_missing(data.frame(a = c(1,2), b = c(NA, 3)))

}
